Support multiple iterations of zigbee2mqtt driver
zigbee2mqtt is in a state of flux, so this accounts for multiple versions of the driver.

I released with the messages as action:button1

My PR went live as action:button_1, etc

Then a subsequent PR by another developer changed the actions to more descriptive names and on the click attribute.
Like click:power, click:brightness. 

This code accounts for all three possibilities and maps to the button number.